Move attachFormOnSubmit into separate file
